+++ python/sglang/README.md
@@ -12,6 +12,5 @@
 - `global_config.py`: The global configs and constants.
 - `launch_server.py`: The entry point for launching the local server.
 - `llama3_eval.py`: Evaluation of Llama 3 using the Meta Llama dataset.
-- `profiler.py`: Profile a running server.
 - `utils.py`: Common utilities.
 - `version.py`: Version info.
