--- python/sglang/bench_latency.py
@@ -1,1 +1,0 @@
-raise ValueError("bench_latency.py has been renamed to bench_one_batch.py")
